requirements.txt
langchain
langchain-google-genai
google-generativeai
python-dotenv
requests
# Use modern MSAL and PyJWT to avoid legacy crypto extras and ensure wheel availability
msal==1.28.0
PyJWT==2.8.0
google-generativeai>=0.8.0
langchain>=0.2.16
langchain-google-genai>=1.0.9
